refactor: add entity package
