+++ src/main/java/hibernate/entity/EntityField.java
@@ -1,4 +1,4 @@
-package hibernate;
+package hibernate.entity;
 
 import jakarta.persistence.Column;
