+++ src/test/java/hibernate/entity/ColumnTypeTest.java
@@ -1,5 +1,6 @@
-package hibernate;
+package hibernate.entity;
 
+import hibernate.entity.ColumnType;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
